LT-2507: pr fixes
